chore(Features): fix broken type_conversion
