+++ src/config/deserialization/config_options.rs
@@ -3,6 +3,8 @@ use eyre::Context;
 use CheckerTool::*;
 
 mod features;
+pub use self::features::Features;
+
 mod type_conversion;
 
 #[derive(Serialize, Deserialize, JsonSchema, Clone)]
